requirements.txt
google-generativeai>=0.8.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
rich>=13.0.0
